ai-agent/requirements.txt

azure-ai-projects
azure-identity
dotenv
streamlit